aladdin: a diamond in the rough next-gen web connectivity
This is an experiment to explore how specific subsets of the next
generation web connectivity nettest would look like.

The probe-engine/miniooni platform already contains enough functionality
to allow us to implement most of the rest as a bash script for now.

Of course, the final objective is to get this right and rewrite all
this in golang, to be integrated in probe-engine.