domain-check.bash: be honest and say maybe more often
To say yes/no we need to parse more deeply the results. This is not
the case currently. But this is fine, since at this stage we want to
collect data to analyze and start sketching out what tests we want
to run, so the real output is fine even if not 100% correct.

Also, there is a disclaimer saying that. Adjust the disclaimer to
clarify that (1) the code is alpha and (2) the logic is alpha. That
is, these are two separate issues, not just one issue.